requirements.txt
streamlit 
PyPDF2
python-docx
langchain 
google-generativeai
dotenv
langchain-community 
langchain-google-genai